Removing license
The moment you need to change or delete the license text, the `remove` command comes into play.
`remove` uses exactly the same flags as `inject`, but unlike `inject`, it deletes the license from all files
determined by `.licignore`. Because GoLic has no state, the `remove` command deletes based on the criteria
specified in the configuration and `.licignore`, so the configuration of deleted licenses must be the same as
the configuration that injected them.
```shell
golic remove --dry -t apache2
```