possibility to generate a license under wildchart expression
resolves #4
I have extended the configuration with the possibility to generate licenses under wildchart expression.
This way I can generate a license in shell scripts under `#!/bin/*`, in go files under `package *` etc...
If nothing is confused , the license is generated as a header file.

I also changed the year from 2021 to 2022

Signed-off-by: kuritka <kuritka@gmail.com>