remove replace directive
- https://github.com/golang/go/issues/40276
I can't use the "replace" directive in the go.mod file. If I use it, the `go install`
command does not run (see discussion in the link above). Therefore I am trying to
achieve the same result using a different go.mod directive
(the mikeDNS library is used in Viper, which we don't import).

I plan to delete the last release and re-release again

Signed-off-by: kuritka <kuritka@gmail.com>